vente/requirements.txt
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
crispy-bootstrap4==2022.1
cryptography==38.0.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.7.0
dj-rest-auth==6.0.0
Django==4.2.5
django-allauth==0.63.6
django-cors-headers==4.4.0
django-crispy-forms>=2.3
django-filter==24.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.1.0
fasteners==0.14.1
fonttools==4.53.1
future==0.18.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.5
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.0.1
matplotlib==3.9.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0  # Vérifiez la compatibilité avec Python 3.12
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
pillow==10.4.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-debian==0.1.49  # Mise à jour avec une version disponible
python3-openid==3.2.0
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2024.1
xdg==5
zipp==1.0.0


